feat(agent) agent logger logback
