+++ femas-agent/femas-agent-core/src/main/java/com/tencent/tsf/femas/agent/config/InterceptPlugin.java
@@ -19,6 +19,7 @@ package com.tencent.tsf.femas.agent.config;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.tencent.tsf.femas.agent.exception.InterceptorWrapperException;
+import com.tencent.tsf.femas.agent.tools.AbstractAgentLogger;
 import com.tencent.tsf.femas.agent.tools.AgentLogger;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.matcher.ElementMatcher;
@@ -33,7 +34,7 @@ import static net.bytebuddy.matcher.ElementMatchers.*;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class InterceptPlugin {
-    private static final AgentLogger LOG = AgentLogger.getLogger(InterceptPlugin.class);
+    private static final AbstractAgentLogger LOG = AgentLogger.getLogger(InterceptPlugin.class);
 
     /**
      * 类名
