+++ femas-agent/femas-agent-core/src/main/java/com/tencent/tsf/femas/agent/interceptor/wrapper/StaticMethodsInterceptOverrideArgsWrapper.java
@@ -21,6 +21,7 @@ package com.tencent.tsf.femas.agent.interceptor.wrapper;
 
 import com.tencent.tsf.femas.agent.classloader.InterceptorClassLoaderCache;
 import com.tencent.tsf.femas.agent.interceptor.StaticMethodsAroundInterceptor;
+import com.tencent.tsf.femas.agent.tools.AbstractAgentLogger;
 import com.tencent.tsf.femas.agent.tools.AgentLogger;
 import net.bytebuddy.implementation.bind.annotation.AllArguments;
 import net.bytebuddy.implementation.bind.annotation.Morph;
@@ -37,7 +38,7 @@ import java.lang.reflect.Method;
 public class StaticMethodsInterceptOverrideArgsWrapper {
 
 
-    private static final AgentLogger LOG = AgentLogger.getLogger(StaticMethodsInterceptOverrideArgsWrapper.class);
+    private static final AbstractAgentLogger LOG = AgentLogger.getLogger(StaticMethodsInterceptOverrideArgsWrapper.class);
 
     private String interceptorClassName;
