+++ femas-agent/femas-agent-core/src/main/java/com/tencent/tsf/femas/agent/transformer/async/transformer/ExecutorTransformer.java
@@ -1,5 +1,6 @@
 package com.tencent.tsf.femas.agent.transformer.async.transformer;
 
+import com.tencent.tsf.femas.agent.tools.AbstractAgentLogger;
 import com.tencent.tsf.femas.agent.tools.AgentLogger;
 import com.tencent.tsf.femas.agent.transformer.async.ClassInfo;
 import javassist.CannotCompileException;
@@ -18,7 +19,7 @@ import java.util.Set;
  * @Author leoziltong@tencent.com
  */
 public class ExecutorTransformer extends AbstractTransformer {
-    private static final AgentLogger LOG = AgentLogger.getLogger(ExecutorTransformer.class);
+    private static final AbstractAgentLogger LOG = AgentLogger.getLogger(ExecutorTransformer.class);
 
     private static Set<String> EXECUTOR_CLASS_NAMES = new HashSet<>();
     private static final Set<String> INTERCEPT_METHEDS = new HashSet<>();
